scraping/requirements.txt
requests==2.25.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
bs4==0.0.1